Abstract HTTP Client behind HTTPlug.
