+++ composer.json
@@ -1,21 +1,28 @@
 {
-    "name": "codenexus/zipcode",
-    "description": "Zip Code API",
-    "keywords": ["php", "codenexus", "laravel", "zipcode"],
+    "name": "codenexus/zipcode-php-client",
+    "description": "A PHP wrapper for using the Zip Code API.",
+    "homepage": "https://github.com/codenexus/zipcode-php-client",
+    "keywords": ["zip code", "zipcode", "api", "wrapper", "php"],
     "license": "MIT",
     "authors": [
       {
-        "name": "CodeNexus LLC",
-        "email": "help@codenexus.org",
+        "name": "Jeremy Schaffer",
+        "email": "jeremys@codenexus.org",
         "homepage": "https://codenexus.org"
       }
     ],
     "require": {
-        "illuminate/support": "^5.3",
-        "guzzlehttp/guzzle": "^6.2"
+      "php": ">=5.6.4",
+      "php-http/message-factory": "^1.0",
+      "php-http/client-implementation": "^1.0",
+      "php-http/discovery": "^1.0"
+
     },
     "require-dev": {
-        "phpunit/phpunit": "4.*"
+        "phpunit/phpunit": "^5.0",
+        "php-http/mock-client": "^0.3",
+        "php-http/message": "^1.0",
+        "guzzlehttp/psr7": "^1.0"
     },
     "autoload": {
         "psr-4": {
