--- src/Laravel/Facades/Zipcode.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Codenexus\Zipcode\Laravel\Facades;
-
-use Illuminate\Support\Facades\Facade;
-
-class Zipcode extends Facade
-{
-    protected static function getFacadeAccessor()
-    {
-        return 'zipcode';
-    }
-}
